Feat: update Ep types and methods for Type prop
* Previously, the rest methods had a sub hierarchy defining the method
used, such as _get or _post. However, through some experimentation, it
became clear that there was no reason to introduce a control flow for
this.

If the user still needs to know what method they are using, they can
now use Type property that tells them whether they are using a
'get', 'post' or any other method that they are working with. This
change also has the added benefit of
simplifying generic input into the validation and preparation methods

* Tests updated accordingly with the shape change

* validate and prepare functions updated accordingly with the type
change.